refactor: move ProtableText, ULTick and TextBlockCardImageCta
