+++ src/components/Posts/PostPortableText/PostTextAndMedia.tsx
@@ -9,7 +9,7 @@ import CMSImage from "@/components/SharedComponents/CMSImage";
 import Flex from "@/components/SharedComponents/Flex";
 import { Heading } from "@/components/SharedComponents/Typography";
 import VideoPlayer from "@/components/VideoPlayer";
-import { PortableTextWithDefaults } from "@/components/PortableText";
+import { PortableTextWithDefaults } from "@/components/SharedComponents/PortableText";
 
 // When we get the JSON portable text it doesn't have the same field names as
 // our generic types / what comes from our graphql queries
