+++ src/components/SharedComponents/PortableText/PTActionTrigger.tsx
@@ -1,9 +1,8 @@
 import React, { FC } from "react";
 import { PortableTextMarkComponent } from "@portabletext/react";
 
-import UnstyledButton from "../UnstyledButton";
-import { useCookieConsent } from "../../browser-lib/cookie-consent/CookieConsentProvider";
-
+import UnstyledButton from "@/components/UnstyledButton";
+import { useCookieConsent } from "@/browser-lib/cookie-consent/CookieConsentProvider";
 import { Span } from "@/components/SharedComponents/Typography";
 
 enum AllowedActions {
