+++ src/pages/contact-us.tsx
@@ -20,7 +20,7 @@ import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
 import BrushBorders from "@/components/SharedComponents/SpriteSheet/BrushSvgs/BrushBorders";
 import NewsletterFormWrap from "@/components/Forms/NewsletterForm/NewsletterFormWrap";
 import getPageProps from "@/node-lib/getPageProps";
-import { PortableTextWithDefaults } from "@/components/PortableText";
+import { PortableTextWithDefaults } from "@/components/SharedComponents/PortableText";
 
 export type ContactPageProps = {
   pageData: ContactPage;
